Refactored and fixed some unit tests
